Completed the media download feature
Currently it can only download a video of 144p resolution, it can be tweaked to download a specific resolution in the future. As of now being able to download the video is the main feature of this pacakage.

Use cases are shown as well as documentation. However, use cases for the media class is not completed yet, as it is a very long list. It will be completed within a few days.